Layer drawing separation success
