+++ src/LayerSettings.svelte
@@ -1,7 +1,7 @@
 <script context="module">
 
     import { onMount } from 'svelte';
-    import {colors, numBarShow, startingPoint, layerWidth, lineWidth} from './Constants.svelte';
+    import {width, height, colors, numBarShow, startingPoint, layerWidth, lineWidth, layerInstLineWidth, maxAmpRadius} from './Constants.svelte';
 
     export function timeCursorMake(p5, cursorHeight){
         let timeCursor = new p5.Sprite(100, 100, lineWidth*15*2, lineWidth*15*2);
@@ -32,7 +32,86 @@
     export function timeCursorRemove(timeCursor){
         timeCursor.remove()
     }
+            //For element moving
 
+    let showLocation = 0
+    function timeToX(bar, start, showLoc = showLocation){
+        showLocation =showLoc;
+        let tick = (bar-1)*256 +start - showLocation;
+        let X = tick/(numBarShow*256) * layerWidth + startingPoint;
+        return X;
+    }
+
+    export function grid(p5, gridHeight, showLocation){
+        p5.strokeCap(p5.ROUND)
+        p5.strokeWeight(lineWidth);
+        let gridColor = p5.color(colors.default);
+        gridColor.setAlpha(50)
+        p5.stroke(gridColor);
+        for (let i = 0 ; i<=numBarShow; i++){
+            let X = timeToX(Math.floor(showLocation/256)+i+1, 0, showLocation);
+            if (X>startingPoint){p5.line(X, 0, X, gridHeight);}
+        }
+    }
+
+    let layerColor
+    export function layerColoring(inst, p5){
+        
+        if (inst == 'Piano'){layerColor =p5.color(colors.blue)}
+        else if (inst == 'Trumpet'){layerColor =p5.color(colors.purple)}
+        else if (inst == 'Base'){layerColor =p5.color(colors.purple)}
+        else if (inst == 'Snare'){layerColor =p5.color(colors.yellow)}
+        else if (inst == 'Cymbal'){layerColor =p5.color(colors.pink)}
+        else if (inst == 'Guitar'){layerColor =p5.color(colors.green)}
+        layerColor.setAlpha(90);
+        return layerColor
+    }
+
+    export function layerdrawing(p5, yLocation, layer){
+        let points = layer.points;
+        let inst = layer.Inst;
+        p5.strokeWeight(lineWidth);
+        p5.fill(p5.color(colors.default));
+        p5.stroke(p5.color(colors.default));
+        p5.strokeCap(p5.SQUARE)
+        p5.line(startingPoint, yLocation, width, yLocation);
+        p5.noStroke();
+        p5.fill(colors.default);
+        p5.ellipse(startingPoint, yLocation, lineWidth*10);
+        p5.blendMode(p5.HARD_LIGHT);
+
+        layerColor = layerColoring(inst, p5)
+        if (inst == 'Piano' || inst == 'Trumpet'){
+            p5.strokeCap(p5.ROUND);
+            p5.strokeWeight(layerInstLineWidth);
+            p5.stroke(layerColor);
+            for (let point of points){
+                let xi = timeToX(point.bar, point.start);
+                let xf = timeToX(point.bar, point.start+point.duration);
+                //console.log(xi, xf)
+                if (xf>startingPoint && xi<width){
+                    if (xf>width){xf = width}
+                    if (xi<startingPoint){xi = startingPoint}
+                    p5.line(xi, yLocation, xf, yLocation);
+                }
+            }
+            p5.strokeCap(p5.SQUARE);
+        } else if (inst == 'Base' || inst == 'Snare' || inst == 'Cymbal' || inst == 'Guitar'){
+            p5.noStroke();
+            p5.fill(layerColor)
+            for (let point of points){
+                let x = timeToX(point.bar, point.start);
+
+                if (x>startingPoint && x<width){
+                    if (x>width){x = width}
+                    if (x<startingPoint){x = startingPoint}
+                    p5.ellipse(x, yLocation, maxAmpRadius*Math.sqrt(point.amp/100));
+                }
+            }
+            p5.strokeCap(p5.SQUARE);
+        } 
+        p5.blendMode(p5.BLEND);
+    }
